requirements.txt
moviepy==1.0.3
numpy>=1.18.1
imageio>=2.5.0
decorator>=4.3.0
tqdm>=4.0.0
Pillow>=7.0.0
scipy>=1.3.0
pydub>=0.23.0
audiofile>=0.0.0
opencv-python>=4.5
svgwrite